+++ lib/component/music/wavlake/wavlake_track_music_info_builder.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:developer';
 
-import 'package:nostr_sdk/utils/string_util.dart';
+import 'package:nostr_sdk/nostr_sdk.dart';
 import 'package:nostrmo/component/music/music_info_builder.dart';
 import 'package:nostrmo/provider/music_provider.dart';
 import 'package:nostrmo/util/dio_util.dart';