+++ lib/router/notice/notice_list_item_widget.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get_time_ago/get_time_ago.dart';
+import 'package:nostr_sdk/nostr_sdk.dart';
 import 'package:nostrmo/provider/notice_provider.dart';
 
 import '../../consts/base.dart';
-import 'package:nostr_sdk/utils/string_util.dart';
 
 class NoticeListItemWidget extends StatelessWidget {
   NoticeData notice;